fix undefine error

+++ FertilizerTradingApp/DTO/Order.cs
@@ -11,7 +11,7 @@ namespace FertilizerTradingApp.Models
 		public string OrderId { get; set; }
 		public float TotalPrice { get; set; }
 		public DateTime Date { get; set; }
-		public int TotalPayment { get; set; }
+		public float TotalPayment { get; set; }
 		public string CustomerPhone { get; set; }
 		public string AccountId { get; set; }
 		public Order(string orderId, float totalPrice, DateTime date, float totalPayment, string customerPhone, string accountId)
